rename prompts files
