+++ packages/generators/generators/lib/plops/prompts/ct-names-prompts.js
