+++ packages/generators/generators/lib/plops/prompts/default-routes-prompts.js
