+++ packages/generators/generators/lib/plops/prompts/draft-and-publish-prompts.js
