+++ packages/generators/generators/lib/plops/prompts/kind-prompts.js
