+++ packages/generators/generators/lib/templates/collection-type-routes.js.hbs
@@ -6,6 +6,7 @@ module.exports = {
       handler: '{{id}}.find',
       config: {
         policies: [],
+        middlewares: [],
       },
     },
     {
@@ -14,6 +15,7 @@ module.exports = {
       handler: '{{id}}.findOne',
       config: {
         policies: [],
+        middlewares: [],
       },
     },
     {
@@ -22,6 +24,7 @@ module.exports = {
       handler: '{{id}}.create',
       config: {
         policies: [],
+        middlewares: [],
       },
     },
     {
@@ -30,6 +33,7 @@ module.exports = {
       handler: '{{id}}.update',
       config: {
         policies: [],
+        middlewares: [],
       },
     },
     {
@@ -38,6 +42,7 @@ module.exports = {
       handler: '{{id}}.delete',
       config: {
         policies: [],
+        middlewares: [],
       },
     },
   ],
