+++ packages/generators/generators/lib/templates/single-route.js.hbs
@@ -6,6 +6,7 @@ module.exports = {
     //  handler: '{{id}}.exampleAction',
     //  config: {
     //    policies: [],
+    //    middlewares: [],
     //  },
     // },
   ],
