+++ packages/generators/generators/lib/templates/single-type-routes.js.hbs
@@ -6,6 +6,7 @@ module.exports = {
       handler: '{{id}}.find',
       config: {
         policies: [],
+        middlewares: [],
       },
     },
     {
@@ -14,6 +15,7 @@ module.exports = {
       handler: '{{id}}.update',
       config: {
         policies: [],
+        middlewares: [],
       },
     },
     {
@@ -22,6 +24,7 @@ module.exports = {
       handler: '{{id}}.delete',
       config: {
         policies: [],
+        middlewares: [],
       },
     },
   ],
